# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "resources api",
        "title": "commonpool api",
        "termsOfService": "http://swagger.io/terms",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8585",
    "basePath": "/api/v1",
    "paths": {
...
        "/resources": {
            "get": {
                "description": "Search for resources",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "Searches resources",
                "operationId": "searchResources",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search text",
                        "name": "query",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "0",
                            "1"
                        ],
                        "type": "string",
                        "description": "Resource type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Created by",
                        "name": "created_by",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of resources to take",
                        "name": "take",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of resources to skip",
                        "name": "skip",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.SearchResourcesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates a resource",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "Updates a resource",
                "operationId": "updateResource",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Resource id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Resource to create",
                        "name": "resource",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.UpdateResourceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.UpdateResourceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a resource",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "Creates a resource",
                "operationId": "createResource",
                "parameters": [
                    {
                        "description": "Resource to create",
                        "name": "resource",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.CreateResourceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.CreateResourceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    }
                }
            }
        },
        "/resources/:id": {
            "get": {
                "description": "Gets a resource by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "Gets a single resource",
                "operationId": "getResource",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Resource id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.GetResourceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    }
                }
            }
        },
...
        "/users/:id": {
            "get": {
                "description": "Returns information about the given user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Returns information about a user",
                "operationId": "getUserInfo",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "User id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.UserInfoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "errors.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
        "utils.Error": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "web.CreateResourcePayload": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                },
                "valueInHoursFrom": {
                    "type": "integer"
                },
                "valueInHoursTo": {
                    "type": "integer"
                }
            }
        },
        "web.CreateResourceRequest": {
            "type": "object",
            "properties": {
                "resource": {
                    "type": "object",
                    "$ref": "#/definitions/web.CreateResourcePayload"
                }
            }
        },
        "web.CreateResourceResponse": {
            "type": "object",
            "properties": {
                "resource": {
                    "type": "object",
                    "$ref": "#/definitions/web.Resource"
                }
            }
        },
...
        "web.GetResourceResponse": {
            "type": "object",
            "properties": {
                "resource": {
                    "type": "object",
                    "$ref": "#/definitions/web.Resource"
                }
            }
        },
...
        "web.Resource": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "createdById": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                },
                "valueInHoursFrom": {
                    "type": "integer"
                },
                "valueInHoursTo": {
                    "type": "integer"
                }
            }
        },
        "web.SearchResourcesResponse": {
            "type": "object",
            "properties": {
                "resources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.Resource"
                    }
                },
                "skip": {
                    "type": "integer"
                },
                "take": {
                    "type": "integer"
                },
                "totalCount": {
                    "type": "integer"
                }
            }
        },
...
        "web.UpdateResourcePayload": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                },
                "valueInHoursFrom": {
                    "type": "integer"
                },
                "valueInHoursTo": {
                    "type": "integer"
                }
            }
        },
        "web.UpdateResourceRequest": {
            "type": "object",
            "properties": {
                "resource": {
                    "type": "object",
                    "$ref": "#/definitions/web.UpdateResourcePayload"
                }
            }
        },
        "web.UpdateResourceResponse": {
            "type": "object",
            "properties": {
                "resource": {
                    "type": "object",
                    "$ref": "#/definitions/web.Resource"
                }
            }
        },
...
        "web.UserInfoResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}